requirements.txt
chatgptpy==1.0.6
python-telegram-bot==13.15
stability-sdk==0.3.0
Pillow==9.3.0
